% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-cross.R
\name{cluster_em_cross_cor}
\alias{cluster_em_cross_cor}
\title{Cluster Expression-methylation cross-correlation}
\usage{
cluster_em_cross_cor(
  em_list,
  k_meth,
  k_expr = k_meth,
  hc_exp = NULL,
  normal_meth = NULL
)
}
\arguments{
\item{em_list}{output of \code{em_cross_cor}}

\item{k_meth}{number of methylation clusters}

\item{k_expr}{number of expression clusters}

\item{hc_exp}{\code{hclust} object with existing clustering of the matrix columns}
}
\value{
a list with:
\itemize{
\item{"em_cross"}{Matrix with expression-methylation correlation values. Rows are loci and columns are genes}
\item{meth_mat}{Matrix with methylation values (filtered)}
\item{expr_mat}{Matrix with expression values (filtered)}
\item{hc_meth}{hclust object with clustering of the loci}
\item{hc_exp}{hclust object with clustering of the genes}
\item{meth_clust}{tibble with cluster number for each row}
\item{expr_clust}{tibble with cluster number for each column}
\item{expr_mods}{matrix with mean expression in every cluster for each sample}
\item{meth_mods}{matrix with mean methylation in every cluster for each sample}
\item{expr_mods_locus}{matrix with mean expression in every cluster for each locus}
\item{meth_mods_gene}{matrix with mean methylation in every cluster for each gene}
}
}
\description{
Cluster Expression-methylation cross-correlation
}
