% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coord_to_promoter_names}
\alias{coord_to_promoter_names}
\title{Change rownames of matrix from coordinates to promoter names}
\usage{
coord_to_promoter_names(mat, promoter_intervs)
}
\arguments{
\item{mat}{matrix with coordinate rownames (chrom_start_end).}

\item{promoter_intervs}{intervals set with additional "name" column. In case of duplicate names or coordinates the first one in the matrix will be used.}
}
\value{
\code{mat} with rownames changed to promoter name taken from \code{promtoer_intervs}
}
\description{
Change rownames of matrix from coordinates to promoter names
}
