% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME-deconvolution.R
\name{deconv_TME}
\alias{deconv_TME}
\title{Deconvolute TME (tumor microenvironment) effects from methylation data}
\usage{
deconv_TME(
  meth_mat,
  expr_mat,
  raw_meth_mat,
  min_meth = 0.1,
  max_meth = 0.9,
  min_expr = NULL,
  meth_cor_thresh = 0.25,
  expr_cor_thresh = 0.25,
  min_sd = NULL,
  k = NULL,
  k_meth = 30,
  k_expr = k_meth,
  caf_gene = "CAV1",
  immune_gene = "CD3D",
  ...
)
}
\arguments{
\item{meth_mat}{Matrix with methylation values to use for TME inference (usually - promoter methylation). Each row is a locus and each column is a sample. This can be a different matrix than \code{raw_meth_mat}.}

\item{expr_mat}{Matrix with expression values. Each row is a gene and each column is a sample.}

\item{raw_meth_mat}{Matrix with methylation values to deconvolute. Each row is a locus and each column is a sample. This can be a different matrix than \code{meth_mat}.}

\item{min_meth}{minimal locus average methylation}

\item{max_meth}{maximal locus average methylation}

\item{min_expr}{minimal expression level. Only genes with at least one sample with expression above min_expr would be included.}

\item{meth_cor_thresh}{minimal correlation level of a locus. Only loci with at least one correlation above the threshold will be included.}

\item{expr_cor_thresh}{minimal correlation level of a gene. Only genes with at least one correlation above the threshold will be included.}

\item{min_sd}{minimal standard deviation per locus.}

\item{k}{k parameter for knn}

\item{k_meth}{number of methylation clusters when clustering expression-methylation data}

\item{k_expr}{number of expression clusters when clustering expression-methylation data}

\item{caf_gene}{name of gene that is inside the CAF cluster (default: "CAV1")}

\item{immune_gene}{name of gene that is inside the immune cluster (default: "CD3D")}

\item{...}{additional parameters to \code{em_cross_cor}}
}
\value{
a list with:
\itemize{
\item{norm_meth}{Matrix with normalized methylation values of \code{raw_meth_mat}}
\item{tme_features}{Data frame with Immune and CAF expression signatures per sample}
\item{em_cross}{List as returned from \code{em_cross_cor}. Mainly includes a matrix with expression-methylation correlation values that were used for the deconvolution algorithm, where rows are loci and columns are genes}
\item{em_cross_clust}{Clustering of expression-methylation correlation data, as returned from \{code}{cluster_em_cross_cor}}
}
}
\description{
Deconvolute TME (tumor microenvironment) effects from methylation data
}
