% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-cross.R
\name{em_cross_cor}
\alias{em_cross_cor}
\title{Calculate Expression-methylation cross-correlation}
\usage{
em_cross_cor(
  meth_mat,
  expr_mat,
  min_meth = NULL,
  max_meth = NULL,
  min_expr = NULL,
  min_sd = NULL,
  samples = NULL,
  max_na = NULL,
  meth_cor_thresh = NULL,
  expr_cor_thresh = NULL
)
}
\arguments{
\item{meth_mat}{Matrix with methylation values. Each row is a locus and each column is a sample.}

\item{expr_mat}{Matrix with expression values. Each row is a gene and each column is a sample.}

\item{min_meth}{minimal locus average methylation}

\item{max_meth}{maximal locus average methylation}

\item{min_expr}{minimal expression level. Only genes with at least one sample with expression above min_expr would be included.}

\item{min_sd}{minimal standard deviation per locus.}

\item{samples}{names of samples to include. If NULL - all samples that have both expression and methylation data would be included.}

\item{max_na}{maximal number of NAs allowed per locus.}

\item{meth_cor_thresh}{minimal correlation level of a locus. Only loci with at least one correlation above the threshold will be included.}

\item{expr_cor_thresh}{minimal correlation level of a gene. Only genes with at least one correlation above the threshold will be included.}
}
\value{
a list with:
\itemize{
\item{em_cross}{Matrix with expression-methylation correlation values. Rows are loci and columns are genes}
\item{meth_mat}{Matrix with methylation values (filtered)}
\item{expr_mat}{Matrix with expression values (filtered)}
}
}
\description{
Calculate Expression-methylation cross-correlation
}
\examples{

\dontrun{
 init_tcga_brca_example()
 em_list <- get_em_cross(brca_meth_mat, brca_expr_mat, min_meth = 0.1, max_meth = 0.9, meth_cor_thresh = 0.25, expr_cor_thresh = 0.25, min_expr = 2, min_sd = 0.05, max_na = 0)
  

}


}
