% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-cross.R
\name{export_em_cross_mods}
\alias{export_em_cross_mods}
\title{Export Expression-methylation cross-correlation clustering to excel sheet}
\usage{
export_em_cross_mods(em_clust, filename, promoters = NULL)
}
\arguments{
\item{em_clust}{output of \code{cluster_em_cross_cor}}

\item{filename}{name of excel file name}

\item{promoters}{intervals set with additional column 'name' with promoter names}
}
\description{
Export Expression-methylation cross-correlation clustering to excel sheet
}
