% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-em-cross.R
\name{feats_to_knn_matrix}
\alias{feats_to_knn_matrix}
\title{Calculate knn-matrix based on feature matrix}
\usage{
feats_to_knn_matrix(feats, k)
}
\arguments{
\item{k}{k parameter for knn}

\item{feat}{feature matrix (e.g. output of \code{define_immune_cafs_signature})}
}
\value{
knn-matrix: number of samples X number of features with 1 for being knn in the dimesion and 0 otherwise.
}
\description{
Calculate knn-matrix based on feature matrix
}
