% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis-promoters.R
\name{gene_promoter_cors}
\alias{gene_promoter_cors}
\title{Get correlations of all promoters to a specific gene}
\usage{
gene_promoter_cors(
  gene,
  meth_mat,
  expr_mat,
  promoter_intervs,
  spearman = FALSE
)
}
\arguments{
\item{gene}{name of the gene}

\item{meth_mat}{Matrix with methylation values. Each row is a promoter and each column is a sample. Rownames should contain "chrom", "start" and "end" separated by "_"}

\item{expr_mat}{Matrix with expression values. Each row is a gene and each column is a sample.}

\item{promoter_intervs}{intervals set with additional "name" column. In case of duplicate names or coordinates the first one in the matrix will be used. Only intervals that are within promoter_intervs would be used.}

\item{spearman}{use spearman correlation (if FALSE - use pearson)}
}
\value{
dataframe with the following columns:
\itemize{
\item{gene}{name of the gene}
\item{promoter}{name of the promoter}
\item(cor){correlation of the gene}
}
}
\description{
Get correlations of all promoters to a specific gene
}
