% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME-genes.R
\name{get_TME_genes}
\alias{get_TME_genes}
\title{Extract TME genes from methylation expression clustering}
\usage{
get_TME_genes(em_list, immune_gene = "CD3D", caf_gene = "CAV1")
}
\arguments{
\item{em_list}{output of \code{em_cross_cor}}

\item{immune_gene}{name of canonical immune gene (default: CD3D)}

\item{caf_gene}{name of canonical immune gene (default: CAV1)}
}
\value{
vector with names of TME genes
}
\description{
Extracts names of genes that were clustered together with canonical immune and CAF genes in the expression-methylation correlation matrix. This is used mostly for filtering those genes from additional analysis such as cis-regulation.
}
