% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-em-cross.R
\name{immune_cafs_signature_from_clust}
\alias{immune_cafs_signature_from_clust}
\title{Define immune and CAF signatures}
\usage{
immune_cafs_signature_from_clust(
  em_clust,
  scale_feats = TRUE,
  caf_gene = "CAV1",
  immune_gene = "CD3D"
)
}
\arguments{
\item{em_clust}{output of \code{cluster_em_cross_cor}}

\item{caf_gene}{name of gene that is inside the CAF cluster (default: "CAV1")}

\item{immune_gene}{name of gene that is inside the immune cluster (default: "CD3D")}

\item{sacle_feats}{scale features (recommended)}
}
\value{
matrix with 'CAF' and 'immune' score for each sample
}
\description{
Define immune and CAF signatures
}
