% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intervs_to_mat}
\alias{intervs_to_mat}
\title{Transform an intervals set to matrix}
\usage{
intervs_to_mat(df)
}
\arguments{
\item{df}{intervals set}
}
\value{
matrix with coordinate rownames (chrom_start_end).
}
\description{
Transform an intervals set to matrix
}
