% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_to_intervs}
\alias{mat_to_intervs}
\title{Transform a matrix with coordinate rownames to intervals set}
\usage{
mat_to_intervs(mat)
}
\arguments{
\item{mat}{matrix with coordinate rownames (chrom_start_end).}
}
\value{
intervals set
}
\description{
Transform a matrix with coordinate rownames to intervals set
}
