% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylayer-package.R
\docType{package}
\name{methylayer-package}
\alias{methylayer}
\alias{methylayer-package}
\title{methylayer: Layered modeling of tumor methylation}
\description{
Tools for methylation and expression analysis in tumors.
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
