% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-em-cross.R
\name{normalize_immune_cafs_meth}
\alias{normalize_immune_cafs_meth}
\title{Normalize methylation matrix for immune and CAF signatures}
\usage{
normalize_immune_cafs_meth(
  em_clust,
  k,
  meth_mat = em_clust$meth_mat,
  scale_feats = TRUE,
  caf_gene = "CAV1",
  immune_gene = "CD3D"
)
}
\arguments{
\item{em_clust}{output of \code{cluster_em_cross_cor}}

\item{k}{k parameter for knn}

\item{caf_gene}{name of gene that is inside the CAF cluster (default: "CAV1")}

\item{immune_gene}{name of gene that is inside the immune cluster (default: "CD3D")}
}
\value{
normalized methylation matrix
}
\description{
Normalize methylation matrix for immune and CAF signatures
}
