% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-em-cross.R
\name{normalize_mat_knn}
\alias{normalize_mat_knn}
\title{Normalize matrix using knn-matrix}
\usage{
normalize_mat_knn(mat, knn_mat)
}
\arguments{
\item{mat}{matrix to normalize}

\item{knn_mat}{knn_matrix to normalize with}
}
\value{
normalized matrix
}
\description{
Normalize matrix using knn-matrix
}
