% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-em-cross.R
\name{plot_em_cross_cor}
\alias{plot_em_cross_cor}
\title{Plot clustered Expression-methylation cross-correlation matrix}
\usage{
plot_em_cross_cor(
  em_clust,
  fig_ofn = NULL,
  width = 1000,
  height = 1400,
  k_meth = max(meth_clust$clust),
  k_exp = max(expr_clust$clust),
  km_meth = NULL,
  km_exp = NULL,
  downscale = FALSE,
  zlim = c(-0.6, 0.6),
  colors = c("black", "darkred", "white", "darkblue", "cyan")
)
}
\arguments{
\item{em_clust}{output of \code{cluster_em_cross_cor}}

\item{fig_ofn}{filename of the output figure}

\item{width}{width of the saved figure}

\item{height}{height of the saved figure}

\item{k_meth}{k for \code{cutree} of matrix rows}

\item{k_expr}{k for \code{cutree} of matrix columns}
}
\description{
Plot clustered Expression-methylation cross-correlation matrix
}
