% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffle_each_column}
\alias{shuffle_each_column}
\title{Shuffles each column of a numeric matrix}
\usage{
shuffle_each_column(mat)
}
\arguments{
\item{mat}{matrix to shuffle}
}
\value{
matrix in which all the columns are shuffled (each column is shuffled separately)
}
\description{
Shuffles each column of a numeric matrix
}
