% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temla-package.R
\docType{package}
\name{temla-package}
\alias{temla}
\alias{temla-package}
\title{temla: Tumor Expression Methylation Layerd Analysis}
\description{
Tools for methylation and expression analysis in tumors.
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
