% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis-promoters.R
\name{cis_em_promoters}
\alias{cis_em_promoters}
\title{Get candidates for promtoers that are regulated by methylation in cis}
\usage{
cis_em_promoters(
  meth_mat,
  expr_mat,
  promoter_intervs,
  k_locus_rank = 2,
  min_samples = 100
)
}
\arguments{
\item{meth_mat}{Matrix with methylation values. Each row is a promoter and each column is a sample. Rownames should contain the name of the promoter (gene name).}

\item{expr_mat}{Matrix with expression values. Each row is a gene and each column is a sample.}

\item{promoter_intervs}{intervals set with additional "name" column. In case of duplicate names or coordinates the first one in the matrix will be used.}

\item{k_locus_rank}{rank of the correlation to extract for each promoter. For example, if \code{k_locus_rank = 2} the second best correlation would be extracted for every promoter in the field \code{kth}}

\item{min_samples}{minimal number of samples per gene (both expression and methylation)}
}
\value{
dataframe with the following columns:
\itemize{
\item{name}{name of the gene}
\item{rank}{rank of the correlation of the gene with it's promoter}
\item{cor}{correlation of the gene with it's promoter}
\item{kth}{\code{k_locus_rank}th best correlation of the gene with a promoter}
\item{fdr}{false discovery rate for detecting the cis-raget with rank value <= \code{rank}}
}
}
\description{
Get candidates for promtoers that are regulated by methylation in cis
}
